Inital changes post-clippy
