+++ crates/ruff_linter/src/rules/flake8_simplify/mod.rs
@@ -1,5 +1,6 @@
 //! Rules from [flake8-simplify](https://pypi.org/project/flake8-simplify/).
 pub(crate) mod rules;
+mod helpers;
 
 #[cfg(test)]
 mod tests {
