+++ crates/ruff_linter/src/rules/flake8_simplify/rules/mod.rs
@@ -1,4 +1,3 @@
-pub(crate) use ast_bool_op::*;
 pub(crate) use ast_expr::*;
 pub(crate) use ast_ifexp::*;
 pub(crate) use ast_unary_op::*;
@@ -18,8 +17,13 @@ pub(crate) use split_static_string::*;
 pub(crate) use suppressible_exception::*;
 pub(crate) use yoda_conditions::*;
 pub(crate) use zip_dict_keys_and_values::*;
+pub(crate) use compare_with_tuple::*;
+pub(crate) use duplicate_is_instance_call::*;
+pub(crate) use expr_and_false::*;
+pub(crate) use expr_or_not_expr::*;
+pub(crate) use expr_or_true::*;
+pub(crate) use exrp_and_not_expr::*;
 
-mod ast_bool_op;
 mod ast_expr;
 mod ast_ifexp;
 mod ast_unary_op;
@@ -40,3 +44,9 @@ mod split_static_string;
 mod suppressible_exception;
 mod yoda_conditions;
 mod zip_dict_keys_and_values;
+mod compare_with_tuple;
+mod duplicate_is_instance_call;
+mod expr_and_false;
+mod expr_or_not_expr;
+mod expr_or_true;
+mod exrp_and_not_expr;
